Move some test classes into their own package
