+++ subprojects/instant-execution/src/integTest/groovy/org/gradle/instantexecution/inputs/undeclared/GroovyPluginImplementation.groovy
@@ -14,21 +14,18 @@
  * limitations under the License.
  */
 
-package org.gradle.instantexecution
+package org.gradle.instantexecution.inputs.undeclared
 
-import groovy.transform.CompileStatic
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.test.fixtures.file.TestFile
 
-class InstantExecutionUndeclaredBuildInputsStaticGroovyBuildSrcIntegrationTest extends AbstractInstantExecutionUndeclaredBuildInputsIntegrationTest {
-    @Override
-    void buildLogicApplication() {
-        file("buildSrc/src/main/groovy/SneakyPlugin.groovy") << """
+trait GroovyPluginImplementation {
+    void dynamicGroovyPlugin(TestFile sourceFile) {
+        sourceFile << """
             import ${Project.name}
             import ${Plugin.name}
-            import ${CompileStatic.name}
 
-            @CompileStatic
             class SneakyPlugin implements Plugin<Project> {
                 public void apply(Project project) {
                     // Static method call
@@ -40,7 +37,7 @@ class InstantExecutionUndeclaredBuildInputsStaticGroovyBuildSrcIntegrationTest e
                     println("apply CI2 = " + sys.getProperty("CI2"))
 
                     // Call from closure
-                    def cl = { String p ->
+                    def cl = { p ->
                         println("apply \$p = " + sys.getProperty(p))
                     }
                     cl("CI3")
@@ -53,10 +50,6 @@ class InstantExecutionUndeclaredBuildInputsStaticGroovyBuildSrcIntegrationTest e
                     }
                 }
             }
-        """
-
-        buildFile << """
-            apply plugin: SneakyPlugin
         """
     }
 }
